+++ pabilo_payment_gateway/models/__init__.py
@@ -3,6 +3,7 @@ from . import res_company
 from . import res_config_settings
 from . import account_journal
 from . import pabilo_user_bank
+from . import pabilo_verification
 from . import payment_provider
 from . import payment_transaction
 from . import pos_payment_method